+++ traffic_info/__init__.py
@@ -1,19 +1,45 @@
 """traffic_info module."""
+import datetime
 import logging
+import os
 import sys
 
 import configargparse
 
-from .core import Location, MapScreenshot, send_email
+from selenium.common.exceptions import WebDriverException
+
+from workalendar.registry import registry
+
 from .__version__ import __version__
+from .core import Location, MapScreenshot, send_email
+from .exceptions import NotExecutableError, WebdriverNotFoundError
 from .utils import get_chromedriver_path
 
 
-def run() -> None:
-    """Run traffic info from command line."""
-    logger = logging.getLogger(__name__)
-    logger.setLevel(logging.INFO)
+def check_webdriver_path(webdriver: str = None) -> str:
+    """
+    Check if the webdriver exist and is executable.
 
+    Args:
+        webdriver: The webdriver's path.
+
+    Returns:
+        The verified path.
+
+    """
+    if webdriver is None:
+        webdriver = get_chromedriver_path()
+        if webdriver is None:
+            raise WebdriverNotFoundError(webdriver)
+    if not os.path.isfile(webdriver):
+        raise WebdriverNotFoundError(webdriver)
+    if not os.access(webdriver, os.X_OK):
+        raise NotExecutableError(webdriver)
+    return webdriver
+
+
+def parse_args() -> configargparse.Namespace:
+    """Parse command-line arguments."""
     parser = configargparse.ArgParser()
     parser.add_argument(
         "-c", "--config-file", is_config_file=True, help="Config file path."
@@ -47,7 +73,31 @@ def run() -> None:
     parser.add_argument(
         "-H", "--screenshot_height", type=int, help="Screenshot’s height."
     )
-    options = parser.parse_args()
+    parser.add_argument(
+        "-C",
+        "--country_code",
+        type=str,
+        help="Country code(ISO 3166-1/ISO 3166-2) to avoid notifications on holidays.",
+    )
+
+    return parser.parse_args()
+
+
+def run() -> None:
+    """Run traffic info from command line."""
+    logger = logging.getLogger(__name__)
+    logger.setLevel(logging.INFO)
+
+    options = parse_args()
+    if options.country_code:
+        calendar_class = registry.get_calendar_class(options.country_code.upper())
+        if not calendar_class:
+            logger.error("Invalid country code")
+            sys.exit(1)
+        calendar = calendar_class()
+        if calendar.is_holiday(datetime.datetime.now()):
+            # Enjoy your holiday! :)
+            sys.exit()
 
     location_keys = ["latitude", "longitude", "zoom"]
     screenshot_keys = ["api_key", "webdriver_path", "width", "height"]
@@ -58,19 +108,21 @@ def run() -> None:
     screenshot_params = {
         k: v for k, v in vars(options).items() if v is not None and k in screenshot_keys
     }
-    if not options.webdriver_path:
-        webdriver_path = get_chromedriver_path()
-        if webdriver_path is None:
-            logger.error(
-                "Unable to find chrome driver, "
-                "please provide ChromeDriver's path using --webdriver."
-            )
-            sys.exit(1)
-        screenshot_params["webdriver_path"] = webdriver_path
+    try:
+        screenshot_params["webdriver_path"] = check_webdriver_path(
+            options.webdriver_path
+        )
+    except (NotExecutableError, WebdriverNotFoundError) as exc:
+        logger.error(exc.msg)
+        sys.exit(1)
 
     location = Location(**location_params)
     screenshot = MapScreenshot(**screenshot_params)
-    screenshot.take(location)
+    try:
+        screenshot.take(location)
+    except WebDriverException as exc:
+        logger.error(exc.msg)
+        sys.exit(1)
     send_email(options.email_from, options.email_to, location, screenshot)